requirements.txt
pyserial
bitstring

sqlalchemy
influxdb